Revert "Merging server with client code."
This reverts commit c7f8b2fd21776001775008c6f76582ce55494f78.

+++ src/serializermonitor.cpp
@@ -149,7 +149,7 @@ SerializerMonitorStdv3::serializePlayMode( std::ostream & os,
     static const char * playmode_strings[] = PLAYMODE_STRINGS;
 
     os << "(playmode " << time
-       << ' ' << playmode_strings[static_cast<int>(pmode)]
+       << ' ' << playmode_strings[pmode]
        << ')';
 }
 
@@ -171,7 +171,7 @@ void
 SerializerMonitorStdv3::serializePlayModeId( std::ostream & os,
                                              const PlayMode pmode ) const
 {
-    os << " (pm " << static_cast<int>(pmode) << ')';
+    os << " (pm " << pmode << ')';
 }
 
 void
